Cast Device and Driver in log message to unsigned long
When building Android S this is during the build ending up as
both unsigned long long and unsigned long, this makes the
build break when building the tests.
As both seem to be used during a normal build it will not help
to just change the %ul to %ull as you get an error on that instead.

This fixes so it's always printing with unsigned long.

Signed-off-by: Zingo Andersen <zingo.andersen@arm.com>
Change-Id: I1d1e4b8ab72c563a3b570f30112a7ce076f0a54e